Rewrote RiskAnalysisForm logic
