+++ src/pages/ConsumerPurposeEditPage/components/PurposeEditStep2RiskAnalysis/PurposeEditStep2RiskAnalysis.tsx
@@ -1,10 +1,10 @@
 import React from 'react'
 import type { ActiveStepProps } from '@/hooks/useActiveStep'
-import { RiskAnalysisForm, RiskAnalysisFormSkeleton } from './RiskAnalysisForm'
+import { RiskAnalysisForm, RiskAnalysisFormSkeleton } from './RiskAnalysisForm/RiskAnalysisForm'
 import { useNavigate, useParams } from '@/router'
 import { PurposeQueries } from '@/api/purpose'
 import { NotFoundError } from '@/utils/errors.utils'
-import { RiskAnalysisVersionMismatchDialog } from './RiskAnalysisVersionMismatchDialog'
+import { RiskAnalysisVersionMismatchDialog } from './RiskAnalysisForm'
 
 export const PurposeEditStep2RiskAnalysis: React.FC<ActiveStepProps> = (props) => {
   const { purposeId } = useParams<'SUBSCRIBE_PURPOSE_EDIT'>()
