+++ src/pages/ConsumerPurposeEditPage/components/PurposeEditStep2RiskAnalysis/RiskAnalysisForm/index.ts
@@ -1,0 +1,2 @@
+export * from './RiskAnalysisForm'
+export * from './RiskAnalysisVersionMismatchDialog'
