+++ src/pages/ConsumerPurposeEditPage/types/risk-analysis-form.types.ts
@@ -1,0 +1,5 @@
+import type { FormConfigQuestion } from '@/api/api.generatedTypes'
+
+export type AnswerValue = string | Array<string> | boolean
+export type Questions = Record<string, FormConfigQuestion>
+export type Answers = Record<string, AnswerValue>
